requirements.txt
apache-airflow
apache-airflow-providers-postgres
pandas
psycopg2-binary
python-dotenv
sqlalchemy
requests
openpyxl